+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/ActionGraph.java
@@ -116,7 +116,7 @@ public class ActionGraph extends Graph {
      * trait
      */
     private Map<String, Map<String, Trait>> uninitializedClassTraits;
-
+    
     /**
      * Constructs ActionGraph
      *